wip: #60 moved trpc utils file to new lib folder
