+++ apps/issues/pages/projects/archive.tsx
@@ -3,8 +3,8 @@ import Link from 'next/link';
 
 import Head from '../../components/Head';
 import { ProjectList } from '../../components/ProjectList';
+import { trpc } from '../../lib/trpc';
 import requireLayoutProps from '../../utils/requireLayoutProps';
-import { trpc } from '../../utils/trpc';
 
 function Archive() {
   const { data: projects, isLoading } = trpc.project.getAll.useQuery();
